Recursion
It is possible to find a n-th power much quicker than by making n multiplications! To do this you need to use the following recurrence relations:
a^n=(a^2) ^(n/2) for even n,
a^n=a∗a^(n−1) for odd n.
Implement the algorithm of quick exponentiating using a recursion method.