Math library
You are given real numbers a, b and c, where a ≠ 0. 
Solve the quadratic equation ax^2+bx+c=0 and output all of its roots.
It is guaranteed that the equation always has two roots.
Output the results roots in ascending order. Do not round or format them, the testing system does it automatically.