Selection sort in Java (Hyperskill)
Let's say that an array is max-min sorted if the first element of the array is the maximum element, the second is the minimum, the third is the second maximum and so on. Modify Selection sort such that it can be used for max-min sorting.
Input: the first line contains a number n — the length of an input array. The next line contains n numbers — the elements of the array.
Output: a max-min sorted input array.

Sample Input 1:
5
1 2 3 4 5
Sample Output 1:
5 1 4 2 3