Selection sort in Java (Hyperskill)
At each step, Selection sort finds the smallest (largest) element in an unsorted part of an array and places the found element to the end of a sorted part. So, after k iterations are completed, the first k elements are the smallest (largest) elements of the array in ascending (descending) order. This property allows us to implement a partial sorting algorithm which runs in O(k⋅n) for an array of size n. Your task here is to implement such an algorithm by modifying Selection sort.
Input: the first line contains a number n — the length of an input array. The next line contains n numbers — the elements of the array. The 3rd line contains an integer k.

Output: 
k largest elements of the input array in descending order.