ArrayList (Hyperskill)
Write a program that finds out the closest integer to a given one from an array. If you find several integers with the same distance to the N, you should output all of them in the ascending order. If there are several equal numbers, output them all.
Input: a set of integers and a number N.
Output: some number(s) from the given array.

Sample Input 1:
1 2 4 5
3
Sample Output 1:
2 4