The List interface (Hyperskill)
Implement a method that takes an integer value and two lists of numbers. It must check if the value occurs the same number of times in both sequences.
In the following input example, the first line contains the value, the second line is the first list, the third line is another list.
All numbers are separated by whitespaces.