Dealing with modifiers (Hyperskill)
Implement getPrivateFields method to list the names of private fields declared in the class the object belongs to. Fields inherited from parent classes should be omitted. Field names should be sorted in lexical order.

Here we use List<String> since it's more appropriate for the task. Check out the docs if you're not familiar with it.