The for-loop (Hyperskill)
The input contains four numbers: a,b,c,d.

Output all the integer roots between 0 and 1000 (inclusive) for the equation ax^3+bx^2+cx+d=0 in ascending order.
If the specified interval does not contain the roots of the equation, do not output anything.
Remember, that cubic equation always has 3 roots, meaning it can't have more than 3 integer roots. Keep this in mind in order to optimize the code.