Iterating over arrays (Hyperskill)
Write a program that reads an array of integers and two numbers n and m. The program must check that n and m never occur next to each other (in any order) in the array.

Input data format:
The first line contains the size of an array.
The second line contains elements of the array.
The third line contains two integer numbers n and m.
All numbers in the same line are separated by the space character.

Output data format:
The result is a single boolean value true if n and m never occur next to each other; otherwise, it is false.