Dynamic array in Java (Hyperskill)
There is a linear list. The following commands are given:
*add value – add an element with value ‘value’ at the end of the list;
*remove idx – remove the element with index ‘idx’;
*size – return the number of elements in the array;
*isempty – determine whether the list is empty or not. Print «true» or «false»;
*contains value – determine if there is at least one element with value ‘value’ in the array. Print «true» or «false»;
*indexof value – print the index of the first instance of an element with value ‘value’, '-1' otherwise;
*clear – clear the array.

Input: In the first line N is the total number of commands. 0 ≤ N ≤ 100. In the following N are the aforementioned commands. -1000 ≤ value ≤ 1000. It is guaranteed that remove command will not be called for a non-existing index.

Output: For each command size, isempty, contains and indexof print the answer on a separate line.