Dynamic array in Java (Hyperskill)
There is a linear list. The following commands are given:
*add value – add an element with the value ‘value’ at the end of the list;
*remove – remove the last element;
*size – return the number of elements in the array.
Input: In the first line N is the total number of commands. 0 ≤ N ≤ 100. In the following N are aforementioned commands. -1000 ≤ value ≤ 1000. It is guaranteed that remove command will not be called for an empty array.

Output: for each size command on new lines print the answer.