Dynamic array in Java (Hyperskill)
There is a linear list. The following commands are given:
*add value adds an element ‘value’ to the list so that it remains sorted;
*remove idx removes an element with an index 'idx';
*size returns the number of elements in the array;
*indexof value prints the index of the first instance of an element with the value ‘value’, '-1' otherwise.

Input: In the first line N is the total number of commands. 0 ≤ N ≤ 100. In the following N are the aforementioned commands. -1000 ≤ value ≤ 1000. It is guaranteed that remove command will not be called for a non-existing index.

Output: For each command size or indexof print the answer on a separate line.