Doubly linked list in Java (Hyperskill)
You receive whole numbers with absolute value not greater than 1000. You are to construct a list from them. The algorithm for creating the list is: first two elements are joined sequentially. The rest will be added either at the beginning or the end of the list: if the value is closer to the first element than the last one, it is added at the beginning, otherwise at the end. After inserting all the elements, print the resulting list. If the difference between element and head is equal to the difference between element and tail, then add it to the end of the list.

Input: in the first line is N, the number of elements. They are written in the next line divided by spaces.
Output: in the first line print the resulting list divided by spaces.

Sample Input 1:
5
4 2 1 5 3
Sample Output 1:
5 4 2 1 3