Doubly linked list in Java (Hyperskill)
There are 5 types of commands given: add an element to the beginning and to the end, remove from the beginning and the end, and reverse the list. Your task is to print the resulting list.
Input: In the first line, N < 1000 is the number of commands. In the following N lines, there are commands:
addFirst x;
addLast x;
removeFirst;
removeLast;
reverse.
It is guaranteed that removal operation is not called for an empty list.

Output: in the first line print the resulting list divided by spaces.

Sample Input 1:
5
addLast 1
addLast 2
addLast 3
reverse
removeFirst
Sample Output 1:
2 1