StringBuilder (Hyperskill)
You’re given a string with no more than 1000 Latin characters and the commands of two types:
cut the string by given index and merge it back together in a different order. For example, given the index is 2 and the string is "example", first it is broken into "ex" and "ample", and then glued back again as "ampleex";
reverse the whole string. If we reverse "example", we get "elpmaxe".
These commands may be performed multiple times. Your task is to print the resulting string after performing all operations.

Input: The first line is the given string. The second line is N – the number of commands. In the following N lines are commands.
split index – split the string by index ‘index’ and invert the pieces;
reverse – reverse the string.
Output: print the resulting string after all operations.