Trees in Java (Hyperskill)
Given a tree in the format of "parent-child" pairs. Write a program that prints the leaves of the tree in ascending order.

Input: the first line contains an integer n — the number of edges in a tree. Each of the next n lines contains two integers separated by space. Each pair of integers corresponds to an edge of a tree: the first integer is a parent, the second one is a child. The node numbered 0 is a root. It is guaranteed that the input corresponds to a correct tree.

Output: the first line should contain the number of leaves in the tree. The second line should be a sequence of these leaves in ascending order.

Sample Input 1:
3
0 1
0 2
2 5
Sample Output 1:
2
1 5