Trees in Java (Hyperskill)
Given a tree represented as a list of pairs of adjacent nodes. Write a program that checks whether the given tree is a full binary tree.
Input: the first line contains an integer n≤100 — the number of edges in the tree. Each of the following n lines contains a pair of two integers representing an edge. The first integer is a parent node, the second one is a child. All nodes have labels between 0 and n. The node numbered 0 is a root. It is guaranteed that the input corresponds to a correct tree.
Output: print "yes" if the given tree is a full binary and "no" otherwise.