LocalDate (Hyperskill)
Write a program that prints all dates of the given year with a specified offset applied.
It should read a starting date and a value of an offset (in days).
In the output, dates should be printed in ascending order with the starting date included. Do not output the dates from the next year.